fix(search): move encapsulation of SearchTotal to search.js
This ensures that both Dicta and Sefaria queries are using the same format of SearchTotal.
